refactor: create <HomePageOrder> and @/components/Home
