+++ components/Home/HomeCategoryCarousel.tsx
@@ -3,7 +3,7 @@ import ImageFixed from "next/image"
 import Image from "@/components/CustomImage"
 import Link from "@/components/CustomLink"
 
-export default function CategoryCarousel() {
+export default function HomeCategoryCarousel() {
   return (
     <div className="block-class bottom-sm-margin">
       <div className="container mx-auto">
